refactor: split useMessageActions into one composable per action group
